V1.2

+++ Task_7/Program.cs
@@ -7,9 +7,9 @@ namespace Task_7
         static void Main(string[] args)
         {
             Console.WriteLine("How much gold do you have?");
-            int coins = Convert.ToInt32(Console.ReadLine());
+            int coins = Convert.ToInt16(Console.ReadLine());
             Console.WriteLine("How many crystals do you wish to purchase?");
-            int numberCrystals = Convert.ToInt32(Console.ReadLine());
+            int numberCrystals = Convert.ToInt16(Console.ReadLine());
             int crystalPrice = 2;
 
             if ((coins / crystalPrice) >= numberCrystals)
